refactor: extract OptionRow.Container layout subcomponent
